Use Gunicorn as a process manager
Uvicorn is an ASGI application runner - I runs a single thread, and does
not manage processes. As such it ships with a work type for Gunicorn,
leveraging Gunicorn's robust process management but using uvicorn to
handle the Django application itself.

So here we have switched across to using this approach. It also fixes the
entry point so that it's does the correct template replacements for the
project package name.