Use poetry export to ensure pinned requirements
The Docker file builds and then installs the application as a package.
This works quite nicely to isolate the application entirely within a
venv. It does however also mean that the Python dependencies are not
fully pinned like they are for local development. The package that
poetry builds uses the same constraints as those within the
`pyproject.toml` _no_ those defined in the lock file. So there can be
some discrepencies.

The Poetry export plugin can be used to export a `requirements.txt` that
_does_ pin dependencies to specific versions. So here we are using it to
create the `requirements.txt` which is then used in the second stage to
install all the dependencies. The final install step of the wheel is
done with the `--no-deps` flag preventing it from trying to install anyt
dependencies itself as they should all already be installed.