Make sure image run has correct options by default
To ensure the container has access to the same network as the
docker-compose services we need to supply it. This is now done by
default. As is setting the correctly database URL when running inside
the container.

The run-image task was refactored a little to make it easily to
default/override these values. The CI action to check the container
was also abler to be simplified a little as result.