Use full checkout for docker image building
We build the docker tag in part from the number of commits on the
current branch. This provides a number value that can only increaase
over time. A useful automated "release" or "version" number. However, it
relies on `git rev-list` which in turn requires _all_ the commits to be
available in order to get an accurate count. The default for github
actions is a shallow fetch. So here we change it to fetch the whole
shebang whem building the Docker image.