Remove Gunicorn
Uvicorn now has process management, so we can use that directly and
remove the need for Gunicorn.

I also took the opportunity to update the default logging so it pretty
closely matches the default logging format used by uvicorn. This
prevents the jarring changed when the Django logging is configured.